move CrateIndex into its own module
